ENG-01: extract _limits (T2/SPE/score limits, ellipse_coordinates)
Move the confidence-limit and ellipse-geometry helpers into _limits and
drop the now-unused scipy.stats chi2/f/norm import from _pca_pls (t_dist
stays for PLS). _pca_pls re-imports the moved names.
